requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
facenet-pytorch==2.5.2
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
urllib3==1.26.6
